backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

mysql-connector-python==8.2.0
PyMySQL==1.1.0

python-docx==1.1.0
PyMuPDF==1.23.8
reportlab==4.0.7
pytesseract==0.3.10
Pillow==10.2.0

torch --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu

sentence-transformers==2.3.1
faiss-cpu==1.7.3

numpy==1.26.4
scipy==1.10.1
nltk==3.8.1
requests==2.31.0

# For Qwen3 chatbot (requires transformers>=4.51.0)
transformers>=4.51.0
accelerate==0.24.1
safetensors==0.4.1
